+++ src/components/header/header.spec.jsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import raf from '../../../test/shims/react-raf-shim';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
