+++ src/components/slack-icon/slack-icon.jsx
