+++ src/components/twitter-icon/twitter-icon.spec.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import { shallow, configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+import TwitterIcon from './twitter-icon';
+
+configure({ adapter: new Adapter() });
+
+describe('Slack-Icon SVG Component', () => {
+  const icon = shallow(<TwitterIcon />);
+
+  it('should not be null', () => {
+    expect(icon).not.toBeNull();
+  });
+});
